+++ plugin_qgis_lpo/processing/provider.py
@@ -30,7 +30,6 @@ from qgis.PyQt.QtGui import QIcon
 
 from .extract_data import ExtractData
 from .extract_data_observers import ExtractDataObservers
-from .extract_data_old import ExtractDataOld
 from .state_of_knowledge import StateOfKnowledge
 from .summary_map import SummaryMap
 from .summary_table_per_species import SummaryTablePerSpecies
@@ -68,9 +67,6 @@ class Provider(QgsProcessingProvider):
         Should return a QIcon which is used for your provider inside
         the Processing toolbox.
         """
-        icon = QIcon(
-            os.path.join(plugin_path, os.pardir, "icons", "logo_lpo_aura_carre.png")
-        )
         return QIcon(
             os.path.join(plugin_path, os.pardir, "icons", "logo_lpo_aura_carre.png")
         )
@@ -93,7 +89,6 @@ class Provider(QgsProcessingProvider):
             SummaryTablePerTimeInterval(),
             StateOfKnowledge(),
             SummaryMap(),
-            ExtractDataOld(),
         ]
         for alg in algorithms:
             self.addAlgorithm(alg)
